requirements.txt
pysc2 == 2.0.2
absl-py == 0.7.0